# feat: performance optimization for tab navigation