--- a/requirements-dev.txt
+++ b/requirements-dev.txt
@@ -1,0 +1,18 @@
+# --- Python testing ---
+pytest~=8.2
+pytest-asyncio~=0.23
+faker~=25.0           # data generation
+freezegun~=1.4        # freeze time
+testcontainers[postgres]~=4.5
+schemathesis~=3.26    # OpenAPI contract tests
+httpx~=0.27
+locust~=2.26
+black==24.3.0
+ruff==0.4.4
+bandit==1.7.8
+pip-audit==2.7.3
+
+# --- CrewAI specific testing ---
+crewai~=0.1.32
+langchain~=0.1.0
+openai>=1.7.1